Refactor configuring external service clients.
